refactor(lox-space/lox-units): move unit wrappers to lox-space and add feature
