+++ Cargo.toml
@@ -27,7 +27,7 @@ csv = "1.3.1"
 fast_polynomial = "0.3.0"
 float_eq = "1.0.1"
 glam = "0.30.6"
-hashbrown = {version = "0.16.0", features = ["rayon"]}
+hashbrown = { version = "0.16.0", features = ["rayon"] }
 itertools = "0.14.0"
 libm = "0.2.15"
 nom = "8.0.0"
@@ -36,7 +36,7 @@ num-derive = "0.4.2"
 num-traits = "0.2.19"
 numpy = "0.26.0"
 proptest = "1.7.0"
-pyo3 = {version = "0.26.0", features = ["hashbrown"]}
+pyo3 = { version = "0.26.0", features = ["hashbrown"] }
 quick-xml = { version = "0.38.3", features = ["serde", "serialize"] }
 rayon = "1.11.0"
 regex = "1.11.2"
@@ -45,4 +45,4 @@ serde = { version = "1.0.225", features = ["derive"] }
 serde-aux = "4.7.0"
 serde_json = "1.0.145"
 sgp4 = "2.3.0"
-thiserror = "2.0.16"
+thiserror = { version = "2.0.16", default-features = false }
