+++ crates/lox-io/Cargo.toml
@@ -19,7 +19,7 @@ regex.workspace = true
 serde-aux.workspace = true
 serde.workspace = true
 serde_json.workspace = true
-thiserror.workspace = true
+thiserror = { workspace = true, features = ["std"] }
 
 [dev-dependencies]
 rstest.workspace = true
