+++ crates/lox-space/pyproject.toml
@@ -17,7 +17,7 @@ requires = ["maturin>=1.3,<2.0"]
 build-backend = "maturin"
 
 [tool.maturin]
-features = ["pyo3/extension-module"]
+features = ["pyo3/extension-module", "python"]
 
 [tool.pytest.ini_options]
 markers = [
