+++ crates/lox-units/src/lib.rs
@@ -1,13 +1,7 @@
 #![no_std]
 
-#[cfg(feature = "std")]
-extern crate std;
-
 pub mod constants;
 pub mod coords;
 pub mod units;
-
-#[cfg(feature = "python")]
-pub mod python;
 
 pub use crate::units::*;
